+++ pkg/fleet/installer/packages/BUILD.bazel
@@ -65,8 +65,8 @@ go_library(
         ],
         "@rules_go//go/platform:windows": [
             "//pkg/fleet/installer/msi",
-            "//pkg/fleet/installer/packages/embedded",
             "//pkg/fleet/installer/packages/exec",
+            "//pkg/fleet/installer/packages/processmanager",
             "//pkg/fleet/installer/packages/service/windows",
             "//pkg/fleet/installer/packages/ssi",
             "//pkg/fleet/installer/packages/user/windows",
