+++ pkg/fleet/installer/packages/processmanager/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "processmanager",
+    srcs = [
+        "ddot_procmgr_config_windows.go",
+        "enabled_windows.go",
+        "reload_windows.go",
+    ],
+    importpath = "github.com/DataDog/datadog-agent/pkg/fleet/installer/packages/processmanager",
+    target_compatible_with = ["@platforms//os:windows"],
+    visibility = ["//visibility:public"],
+    deps = select({
+        "@rules_go//go/platform:windows": [
+            "//pkg/fleet/installer/packages/embedded",
+            "//pkg/fleet/installer/paths",
+            "//pkg/util/log",
+            "//pkg/util/winutil",
+            "@in_yaml_go_yaml_v2//:yaml",
+            "@org_golang_x_sys//windows/registry",
+        ],
+        "//conditions:default": [],
+    }),
+)
